move new constructor to dtypes/dtypes.py
