+++ pandas/core/arrays/categorical.py
@@ -200,71 +200,6 @@ def contains(cat, key, container):
         return any(loc_ in container for loc_ in loc)
 
 
-def create_categorical_dtype(values=None, categories=None, ordered=None,
-                             dtype=None):
-    """
-    Construct and return a :class:`~pandas.api.types.CategoricalDtype`.
-
-    This is a helper function, and specifically does not do the
-    factorization step, if that is needed.
-
-    Parameters
-    ----------
-    values : list-like, optional
-        The list-like must be 1-dimensional.
-    categories : list-like, optional
-        Categories for the CategoricalDtype.
-    ordered : bool, optional
-        Designating if the categories are ordered.
-    dtype : CategoricalDtype or the string "category", optional
-        If ``CategoricalDtype`` cannot be used together with
-        `categories` or `ordered`.
-
-    Returns
-    -------
-    CategoricalDtype
-
-    Examples
-    --------
-    >>> create_categorical_dtype()
-    CategoricalDtype(categories=None, ordered=None)
-    >>> create_categorical_dtype(categories=['a', 'b'], ordered=True)
-    CategoricalDtype(categories=['a', 'b'], ordered=True)
-    >>> dtype1 = CategoricalDtype(['a', 'b'], ordered=True)
-    >>> dtype2 = CategoricalDtype(['x', 'y'], ordered=False)
-    >>> c = Categorical([0, 1], dtype=dtype1, fastpath=True)
-    >>> create_categorical_dtype(c, ['x', 'y'], ordered=True, dtype=dtype2)
-    ValueError: Cannot specify `categories` or `ordered` together with `dtype`.
-
-    The supplied dtype takes precedence over values's dtype:
-
-    >>> create_categorical_dtype(c, dtype=dtype2)
-    CategoricalDtype(['x', 'y'], ordered=False)
-    """
-    if dtype is not None:
-        # The dtype argument takes precedence over values.dtype (if any)
-        if isinstance(dtype, compat.string_types):
-            if dtype == 'category':
-                dtype = CategoricalDtype(categories, ordered)
-            else:
-                msg = "Unknown dtype {dtype!r}"
-                raise ValueError(msg.format(dtype=dtype))
-        elif categories is not None or ordered is not None:
-            raise ValueError("Cannot specify `categories` or `ordered` "
-                             "together with `dtype`.")
-    elif is_categorical(values):
-        # If no "dtype" was passed, use the one from "values", but honor
-        # the "ordered" and "categories" arguments
-        dtype = values.dtype._from_categorical_dtype(values.dtype,
-                                                     categories, ordered)
-    else:
-        # If dtype=None and values is not categorical, create a new dtype.
-        # Note: This could potentially have categories=None and ordered=None.
-        dtype = CategoricalDtype(categories, ordered)
-
-    return dtype
-
-
 _codes_doc = """\
 The category codes of this categorical.
 
@@ -381,7 +316,8 @@ class Categorical(ExtensionArray, PandasObject):
     def __init__(self, values, categories=None, ordered=None, dtype=None,
                  fastpath=False):
 
-        dtype = create_categorical_dtype(values, categories, ordered, dtype)
+        dtype = CategoricalDtype._from_values_or_dtype(values, categories,
+                                                       ordered, dtype)
         # At this point, dtype is always a CategoricalDtype, but
         # we may have dtype.categories be None, and we need to
         # infer categories in a factorization step futher below
@@ -689,7 +625,8 @@ class Categorical(ExtensionArray, PandasObject):
             categorical. If not given, the resulting categorical will be
             unordered.
         """
-        dtype = create_categorical_dtype(codes, categories, ordered)
+        dtype = CategoricalDtype._from_values_or_dtype(codes, categories,
+                                                       ordered)
 
         codes = np.asarray(codes)  # #21767
         if not is_integer_dtype(codes):
